Create Class117.py
Modified my original code that followed the pltw notebook on clever into an organized code with a class called TurtleSpiral. Instead of mainly updating the for loop, modified variables and defined them in a class, and used string function to print out values, just like step 6 on the pltw notebook.